fix: add stricter specifier for sat types
